Extract jsonb/hstore common query code into ActiveRecord::PgQueryMethods
